Move temba tag tests all into same place
